Fix realsense camera latency
First, we need to start the camera early, because there is
a latency spike at the first `.wait_for_frames()` call, it
takes about 500ms to consume the frist frame, so we start it
early to mitigate this issue. The down side is that we will
spend more time on hardware initialization, but that's fine.

Second, now we use `threading.Event()` to tell when the
whole capturing started, in that way, capture thread
can start at the most possible time point.

For the camera latency, in my case, using `rs-latency-tool`
I get an average 90 ms delay on my realsense camera,
that means on a 30 FPS capture, I want to defer 3 frames then
start to write the color image to the file.

Illustrating it as a figure would be like this:

       -90ms       -60ms      -30ms       0ms
 ---------------------------------------------------> Time
         |          |          |  <mmwave>| `.get_adc()`   -> mmwave
                                   <cam-0>| `.get_frame()` -> ideal
  <cam-0>|                                | `.get_frame()  -> cam-0
                                             (call at 0 ms)